refactor: 베팅 금액 타입을 int 에서 long으로 변경

+++ src/main/java/domain/result/PlayersMatchResult.java
@@ -8,10 +8,10 @@ import java.util.Map;
 public record PlayersMatchResult(Map<Player, MatchCase> playerMatchResult) {
 
     public PlayersBettingProfit calculateBettingProfit() {
-        Map<Player, Integer> bettingResult = new LinkedHashMap<>();
+        Map<Player, Long> bettingResult = new LinkedHashMap<>();
 
         for (Player player : playerMatchResult.keySet()) {
-            int profit = player.calculateBettingProfit(playerMatchResult.get(player));
+            long profit = player.calculateBettingProfit(playerMatchResult.get(player));
             bettingResult.put(player, profit);
         }
         return new PlayersBettingProfit(bettingResult);
